+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTablePanelPreferencesTest.java
@@ -68,7 +68,7 @@ public class EntityTablePanelPreferencesTest {
 	@AfterEach
 	void tearDown() throws BackingStoreException {
 		preferences.clear();
-		UserPreferences.flush();
+		preferences.flush();
 	}
 
 	@Test